# alphametics: improve performance of example script