Ref issue #78 

This version is translated from the ["compact" Ruby reference implementation](https://github.com/exercism/ruby/blob/master/exercises/alphametics/.meta/solutions/alphametics_compact.rb).


Timing of current tests (with "msec" added to tcltest "-verbose" config)
```
$ time tclsh alphametics.test
++++ alphametics-1 took 4 ms
++++ alphametics-2 took 0 ms
++++ alphametics-3 took 26 ms
++++ alphametics-4 took 5 ms
++++ alphametics-5 took 23 ms
++++ alphametics-6 took 289 ms
++++ alphametics-7 took 2189 ms
++++ alphametics-8 took 9865 ms
++++ alphametics-9 took 24579 ms
++++ alphametics-10 took 29808 ms
alphametics.test:	Total	10	Passed	10	Skipped	0	Failed	0

________________________________________________________
Executed in   66.92 secs   fish           external
   usr time   66.02 secs  459.00 micros   66.02 secs
   sys time    0.76 secs  1390.00 micros    0.76 secs
```

Timing of this new version
```
$ time tclsh alphametics.test
++++ alphametics-1 took 4 ms
++++ alphametics-2 took 1 ms
++++ alphametics-3 took 1 ms
++++ alphametics-4 took 3 ms
++++ alphametics-5 took 4 ms
++++ alphametics-6 took 36 ms
++++ alphametics-7 took 256 ms
++++ alphametics-8 took 579 ms
++++ alphametics-9 took 1389 ms
++++ alphametics-10 took 17658 ms
alphametics.test:	Total	10	Passed	10	Skipped	0	Failed	0

________________________________________________________
Executed in   20.02 secs   fish           external
   usr time   19.74 secs  361.00 micros   19.74 secs
   sys time    0.27 secs  1234.00 micros    0.26 secs
```