@glennj 2020-05-13
For comparison, this is ruby:
```
$ time ruby alphametics_test.rb
Run options: --seed 2766

# Running:

.........

Finished in 2.087878s, 4.3106 runs/s, 4.3106 assertions/s.

9 runs, 9 assertions, 0 failures, 0 errors, 0 skips

________________________________________________________
Executed in    2.21 secs   fish           external
   usr time    2.17 secs   91.00 micros    2.17 secs
   sys time    0.04 secs  408.00 micros    0.04 secs
```
Clearly an advantage to have the permutation generation done in C.