@glennj 2020-05-14
Comparing to ruby:
```
$ time ruby alphametics_test.rb
Run options: --seed 54363

# Running:

..........

Finished in 12.529172s, 0.7981 runs/s, 0.7981 assertions/s.

10 runs, 10 assertions, 0 failures, 0 errors, 0 skips

________________________________________________________
Executed in   12.68 secs   fish           external
   usr time   12.48 secs  102.00 micros   12.48 secs
   sys time    0.14 secs  510.00 micros    0.14 secs
```